pyproject.toml
[project]
name = "tic-tac-toe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "types-tqdm>=4.67.0.20250809",
]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing --no-cov-on-fail --cov-fail-under=90"